services/scanner-service/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.21
asyncpg==0.28.0
requests==2.31.0
pydantic-settings>=2.2.0
python-dotenv==1.0.0
httpx==0.24.1
pytest==7.4.3
pytest-asyncio==0.21.1